RV64GC Progress
*** WIP ***

Remove generation of absolute value in control module, move to usage location.

ALU.si use absolute value for division if required.

CPU.si use new IL2F IL2D FD2I FD2L F2D D2F integer/float conversions. Extend DMA engine to use new RGBA and ARGB pixelblocks.

cpu_functionblocks.si correct errors in compressed instruction expansion. ADDIW ensure sign is extended. Correct offset in LDSP SDSP LD SD.

FPU64.si correct only 32bit integer being passed to single cycle. Correct FMV operations. Replace conversions with intlong2float intlong2double floatdouble2int floatdouble2long.

crt0.s use 64 bit save for registers.

** WIP *** BIOS still cannot enter directories. Only loads from root directory.

+++ SOFTWARE/c/library/PAWSdefinitions.h
@@ -214,10 +214,10 @@ struct sprite_update_flag {
 // FOR 2D SOFTWARE VECTORS
 union Point2D {
     struct {
-        short dx;
-        short dy;
+        int dx;
+        int dy;
     };
-    int packed;
+    long packed;
 };
 
 // FOR SOFTWARE DRAWLISTS
